// src/app/models/result.ts
export class Result {
  place: number;
  swimmer_name: string;
  age: number;
  age_min: number;
  age_max: number;
  gender: string;
  club_code: string;
  distance: number;
  discipline: string;
  msa_id: string;
  course: string;
  final_time: number;
  event_date: Date;
  location: string;
  seconds: number;
}